[hive][csv upload] make INTs BIGINTs

+++ superset/db_engine_specs/hive.py
@@ -106,7 +106,7 @@ class HiveEngineSpec(PrestoEngineSpec):
             """maps tableschema's types to hive types"""
             tableschema_to_hive_types = {
                 "boolean": "BOOLEAN",
-                "integer": "INT",
+                "integer": "BIGINT",
                 "number": "DOUBLE",
                 "string": "STRING",
             }
